Add files via upload
With the above notebook my key focus was to understand the difference in data output or prediction accuracy in this case when the dataset is properly pre processed and when it is not. You will find in the above notebook machine learning model implemented first on improperly cleaned data followed by a thorough pre processing and finally modelling agaian.